refactor and restructure UserInformation service
Reorganized the `UserInformation` service, including renaming from `AuthService` and relocating associated files for better alignment and clarity. Removed unnecessary checks for `global.localStorage` in `LocalStorageService`, simplified logic, and updated references across the codebase to reflect these changes. Additionally, deleted unused test files and replaced manual setups with `vi.stubGlobal` in tests.
