+++ src/app/client-layout.tsx
@@ -12,7 +12,7 @@ import ProvideDependencies from '../shared/contexts/use-inject.context';
 import {
   UserInformationService,
   userInformationServiceProvider
-} from '../data-access/auth-service/user-information.service';
+} from '../data-access/user-information/user-information.service';
 import { LocalStorageService, localStorageServiceProvider } from '../data-access/local-storage/local-storage.service';
 import { useMemo } from 'react';
 import { DatabaseService } from '../data-access/database/database.service';
